Midterm
Creating the intial home page is easy if you can remember where to find the html, css and js for the wesite. Then with the other pages, it's as simple as adding the "../" on the other pages. However I did mess up by not adding the ng-app and the script for angular.js wrong.